feat: Introduce the ExponentialGroupByKey/Mut types
