+++ src/exponential_group/mod.rs
@@ -1,210 +1,10 @@
-use std::slice::{from_raw_parts, from_raw_parts_mut};
-use std::cmp::Ordering::{Less, Greater};
-use std::{fmt, marker};
-use crate::{exponential_search_by, offset_from};
+mod exponential_group;
+mod exponential_group_by;
+mod exponential_group_by_key;
 
-macro_rules! exponential_group_by {
-    (struct $name:ident, $elem:ty, $mkslice:ident) => {
-        impl<'a, T: 'a, P> $name<'a, T, P> {
-            #[inline]
-            pub fn is_empty(&self) -> bool {
-                self.ptr == self.end
-            }
-
-            #[inline]
-            pub fn remainder_len(&self) -> usize {
-                unsafe { offset_from(self.end, self.ptr) }
-            }
-        }
-
-        impl<'a, T: 'a, P> std::iter::Iterator for $name<'a, T, P>
-        where P: FnMut(&T, &T) -> bool,
-        {
-            type Item = $elem;
-
-            fn next(&mut self) -> Option<Self::Item> {
-                if self.is_empty() { return None }
-
-                let first = unsafe { &*self.ptr };
-
-                let len = self.remainder_len();
-                let tail = unsafe { $mkslice(self.ptr.add(1), len - 1) };
-
-                let predicate = |x: &T| if (self.predicate)(first, x) { Less } else { Greater };
-                let index = exponential_search_by(tail, predicate).unwrap_err();
-
-                let left = unsafe { $mkslice(self.ptr, index + 1) };
-                self.ptr = unsafe { self.ptr.add(index + 1) };
-
-                Some(left)
-            }
-
-            fn size_hint(&self) -> (usize, Option<usize>) {
-                if self.is_empty() { return (0, Some(0)) }
-
-                let len = self.remainder_len();
-                (1, Some(len))
-            }
-
-            fn last(mut self) -> Option<Self::Item> {
-                self.next_back()
-            }
-        }
-
-        impl<'a, T: 'a, P> std::iter::DoubleEndedIterator for $name<'a, T, P>
-        where P: FnMut(&T, &T) -> bool,
-        {
-            fn next_back(&mut self) -> Option<Self::Item> {
-                if self.is_empty() { return None }
-
-                let last = unsafe { &*self.end.sub(1) };
-
-                let len = self.remainder_len();
-                let head = unsafe { $mkslice(self.ptr, len - 1) };
-
-                let predicate = |x: &T| if (self.predicate)(last, x) { Greater } else { Less };
-                let index = exponential_search_by(head, predicate).unwrap_err();
-
-                let right = unsafe { $mkslice(self.ptr.add(index), len - index) };
-                self.end = unsafe { self.end.sub(len - index) };
-
-                Some(right)
-            }
-        }
-
-        impl<'a, T: 'a, P> std::iter::FusedIterator for $name<'a, T, P>
-        where P: FnMut(&T, &T) -> bool,
-        { }
-    }
-}
-
-/// An iterator that will reutrn non-overlapping groups in the slice using *exponential search*.
-///
-/// It will not necessarily gives contiguous elements to the predicate function.
-/// The predicate function should implement an order consistent with the sort order of the slice.
-pub struct ExponentialGroupBy<'a, T, P> {
-    ptr: *const T,
-    end: *const T,
-    predicate: P,
-    _phantom: marker::PhantomData<&'a T>,
-}
-
-impl<'a, T: 'a, P> ExponentialGroupBy<'a, T, P>
-where P: FnMut(&T, &T) -> bool,
-{
-    pub fn new(slice: &'a [T], predicate: P) -> Self {
-        ExponentialGroupBy {
-            ptr: slice.as_ptr(),
-            end: unsafe { slice.as_ptr().add(slice.len()) },
-            predicate: predicate,
-            _phantom: marker::PhantomData,
-        }
-    }
-}
-
-impl<'a, T: 'a, P> ExponentialGroupBy<'a, T, P> {
-    /// Returns the remainder of the original slice that is going to be
-    /// returned by the iterator.
-    pub fn remainder(&self) -> &[T] {
-        let len = self.remainder_len();
-        unsafe { from_raw_parts(self.ptr, len) }
-    }
-}
-
-impl<'a, T: 'a + fmt::Debug, P> fmt::Debug for ExponentialGroupBy<'a, T, P> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_struct("ExponentialGroupBy")
-            .field("remainder", &self.remainder())
-            .finish()
-    }
-}
-
-exponential_group_by!{ struct ExponentialGroupBy, &'a [T], from_raw_parts }
-
-/// An iterator that will reutrn non-overlapping *mutable* groups
-/// in the slice using *exponential search*.
-///
-/// It will not necessarily gives contiguous elements to the predicate function.
-/// The predicate function should implement an order consistent with the sort order of the slice.
-pub struct ExponentialGroupByMut<'a, T, P> {
-    ptr: *mut T,
-    end: *mut T,
-    predicate: P,
-    _phantom: marker::PhantomData<&'a T>,
-}
-
-impl<'a, T: 'a, P> ExponentialGroupByMut<'a, T, P>
-where P: FnMut(&T, &T) -> bool,
-{
-    pub fn new(slice: &'a mut [T], predicate: P) -> Self {
-        ExponentialGroupByMut {
-            ptr: slice.as_mut_ptr(),
-            end: unsafe { slice.as_mut_ptr().add(slice.len()) },
-            predicate: predicate,
-            _phantom: marker::PhantomData,
-        }
-    }
-}
-
-impl<'a, T: 'a, P> ExponentialGroupByMut<'a, T, P> {
-    /// Returns the remainder of the original slice that is going to be
-    /// returned by the iterator.
-    pub fn into_remainder(self) -> &'a mut [T] {
-        let len = self.remainder_len();
-        unsafe { from_raw_parts_mut(self.ptr, len) }
-    }
-}
-
-impl<'a, T: 'a + fmt::Debug, P> fmt::Debug for ExponentialGroupByMut<'a, T, P> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let len = self.remainder_len();
-        let remainder = unsafe { from_raw_parts(self.ptr, len) };
-
-        f.debug_struct("ExponentialGroupByMut")
-            .field("remaining", &remainder)
-            .finish()
-    }
-}
-
-exponential_group_by!{ struct ExponentialGroupByMut, &'a mut [T], from_raw_parts_mut }
-
-/// An iterator that will return non-overlapping groups of equal elements, according to
-/// the [`PartialEq::eq`] function in the slice using *exponential search*.
-///
-/// It will not necessarily gives contiguous elements to the predicate function.
-/// The predicate function should implement an order consistent with the sort order of the slice.
-///
-/// [`PartialEq::eq`]: https://doc.rust-lang.org/std/cmp/trait.PartialEq.html#tymethod.eq
-pub struct ExponentialGroup<'a, T: 'a>(ExponentialGroupBy<'a, T, fn(&T, &T) -> bool>);
-
-impl<'a, T: 'a> ExponentialGroup<'a, T>
-where T: PartialEq,
-{
-    pub fn new(slice: &'a [T]) -> ExponentialGroup<'a, T> {
-        ExponentialGroup(ExponentialGroupBy::new(slice, PartialEq::eq))
-    }
-}
-
-group_by_wrapped!{ struct ExponentialGroup, &'a [T] }
-
-/// An iterator that will return non-overlapping *mutable* groups of equal elements, according to
-/// the [`PartialEq::eq`] function in the slice using *exponential search*.
-///
-/// It will not necessarily gives contiguous elements to the predicate function.
-/// The predicate function should implement an order consistent with the sort order of the slice.
-///
-/// [`PartialEq::eq`]: https://doc.rust-lang.org/std/cmp/trait.PartialEq.html#tymethod.eq
-pub struct ExponentialGroupMut<'a, T: 'a>(ExponentialGroupByMut<'a, T, fn(&T, &T) -> bool>);
-
-impl<'a, T: 'a> ExponentialGroupMut<'a, T>
-where T: PartialEq,
-{
-    pub fn new(slice: &'a mut [T]) -> ExponentialGroupMut<'a, T> {
-        ExponentialGroupMut(ExponentialGroupByMut::new(slice, PartialEq::eq))
-    }
-}
-
-group_by_wrapped!{ struct ExponentialGroupMut, &'a mut [T] }
+pub use self::exponential_group::{ExponentialGroup, ExponentialGroupMut};
+pub use self::exponential_group_by::{ExponentialGroupBy, ExponentialGroupByMut};
+pub use self::exponential_group_by_key::{ExponentialGroupByKey, ExponentialGroupByKeyMut};
 
 #[cfg(test)]
 mod tests {
